# Api/Models/Dto/DustCategoryUpdateDto.cs
namespace Juliapos.Portal.ProductApi.Api.Models.Dto
{
    /// <summary>
    /// DTO for updating a dust category
    /// </summary>
    public class DustCategoryUpdateDto
    {
        /// <summary>
        /// Name
        /// </summary>
        public string Name { get; set; }

        /// <summary>
        /// Weight (order)
        /// </summary>
        public int Weight { get; set; }
    }
}