# Api/swagger.v1.json
{
  "swagger": "2.0",
  "info": {
    "title": "Portal Product Api",
    "description": "The Portal Product API allows operations to manage products. v1.0",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/dustcategories": {
      "get": {
        "tags": [
          "DustCategories"
        ],
        "summary": "Get dust categories",
        "operationId": "GetDustCategoriesAsync",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the dust category.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DustCategoryDto"
              }
            }
          },
          "400": {
            "description": "Returned when request can not be completed.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "DustCategories"
        ],
        "summary": "Create a new dust category",
        "operationId": "CreateDustCategoryAsync",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/DustCategoryAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the new dust category.",
            "schema": {
              "$ref": "#/definitions/DustCategoryDto"
            }
          },
          "409": {
            "description": "Returned when there is a conflict with another dust category.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      }
    },
    "/api/v1/dustcategories/{id}": {
      "get": {
        "tags": [
          "DustCategories"
        ],
        "summary": "Get dust category with id as parameter",
        "operationId": "GetDustCategoryByIdAsync",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the dust category.",
            "schema": {
              "$ref": "#/definitions/DustCategoryDto"
            }
          },
          "404": {
            "description": "Returned when the dust category was not found.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "400": {
            "description": "Returned when request can not be completed.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "DustCategories"
        ],
        "summary": "Update a dust category",
        "operationId": "UpdateDustCategoryAsync",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/DustCategoryUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the dust category.",
            "schema": {
              "$ref": "#/definitions/DustCategoryDto"
            }
          },
          "409": {
            "description": "Returned when there is a conflict with another dust category.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "DustCategories"
        ],
        "summary": "Delete a product category",
        "operationId": "DeleteDustCategoryAsync",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the deleted dust category.",
            "schema": {
              "$ref": "#/definitions/DustCategoryDto"
            }
          },
          "409": {
            "description": "Returned when the category is not empty.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      }
    },
...
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product with id as parameter",
        "operationId": "GetProductByIdAsync",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the product.",
            "schema": {
              "$ref": "#/definitions/ProductDto"
            }
          },
          "404": {
            "description": "Returned when the product was not found.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "400": {
            "description": "Returned when request can not be completed.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product",
        "operationId": "UpdateProductAsync",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProductUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the full information about the product.",
            "schema": {
              "$ref": "#/definitions/ProductDto"
            }
          },
          "409": {
            "description": "Returned when there is a conflict with another product.",
            "schema": {
              "$ref": "#/definitions/ErrorResultDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a prouct",
        "operationId": "DeleteProductAsync",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "purgeProduct",
            "description": "",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned with the deleted product.",
            "schema": {
              "$ref": "#/definitions/ProductDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "openid"
            ]
          }
        ]
      }
    }
  },
  "definitions": {
    "DustCategoryAddDto": {
      "description": "DTO for adding a dust category",
      "type": "object",
      "properties": {
        "name": {
          "description": "Name",
          "type": "string"
        },
        "weight": {
          "format": "int32",
          "description": "Weight (order)",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "DustCategoryDto": {
      "description": "DTO for Dust category",
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "description": "Id",
          "type": "string"
        },
        "name": {
          "description": "Name",
          "type": "string"
        },
        "weight": {
          "format": "int32",
          "description": "Weight (order)",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "DustCategoryUpdateDto": {
      "description": "DTO for updating a dust category",
      "type": "object",
      "properties": {
        "name": {
          "description": "Name",
          "type": "string"
        },
        "weight": {
          "format": "int32",
          "description": "Weight (order)",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ErrorResultDto": {
      "type": "object",
      "properties": {
        "statusCode": {
          "$ref": "#/definitions/HttpStatusCode"
        },
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "errorCode": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "HttpStatusCode": {
      "enum": [
        "Continue",
        "SwitchingProtocols",
        "Processing",
        "EarlyHints",
        "OK",
        "Created",
        "Accepted",
        "NonAuthoritativeInformation",
        "NoContent",
        "ResetContent",
        "PartialContent",
        "MultiStatus",
        "AlreadyReported",
        "IMUsed",
        "MultipleChoices",
        "MovedPermanently",
        "Found",
        "SeeOther",
        "NotModified",
        "UseProxy",
        "Unused",
        "TemporaryRedirect",
        "PermanentRedirect",
        "BadRequest",
        "Unauthorized",
        "PaymentRequired",
        "Forbidden",
        "NotFound",
        "MethodNotAllowed",
        "NotAcceptable",
        "ProxyAuthenticationRequired",
        "RequestTimeout",
        "Conflict",
        "Gone",
        "LengthRequired",
        "PreconditionFailed",
        "RequestEntityTooLarge",
        "RequestUriTooLong",
        "UnsupportedMediaType",
        "RequestedRangeNotSatisfiable",
        "ExpectationFailed",
        "MisdirectedRequest",
        "UnprocessableEntity",
        "Locked",
        "FailedDependency",
        "UpgradeRequired",
        "PreconditionRequired",
        "TooManyRequests",
        "RequestHeaderFieldsTooLarge",
        "UnavailableForLegalReasons",
        "InternalServerError",
        "NotImplemented",
        "BadGateway",
        "ServiceUnavailable",
        "GatewayTimeout",
        "HttpVersionNotSupported",
        "VariantAlsoNegotiates",
        "InsufficientStorage",
        "LoopDetected",
        "NotExtended",
        "NetworkAuthenticationRequired"
      ],
      "type": "string"
    },
...
    "ProductDto": {
      "description": "",
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "productCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "dustCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "menuCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "menuName": {
          "type": "string"
        },
        "vatLevel": {
          "$ref": "#/definitions/VatLevel"
        },
        "description": {
          "type": "string"
        },
        "inInventory": {
          "type": "boolean"
        },
        "percentage": {
          "format": "float",
          "type": "number"
        },
        "ascendingStock": {
          "type": "boolean"
        },
        "state": {
          "$ref": "#/definitions/RecordState"
        },
        "created": {
          "format": "date-time",
          "type": "string"
        },
        "userCreate": {
          "type": "string"
        },
        "updated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdate": {
          "type": "string"
        },
        "properties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PropertyReferenceDto"
          }
        },
        "variations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariationAddDto"
          }
        }
      },
      "additionalProperties": false
    },
    "ProductSelectionPageReferenceDto": {
      "description": "",
      "type": "object",
      "properties": {
        "selectionPageId": {
          "format": "uuid",
          "type": "string"
        },
        "foreColor": {
          "type": "string"
        },
        "backColor": {
          "type": "string"
        },
        "rowIdx": {
          "format": "int32",
          "type": "integer"
        },
        "columnIdx": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "ProductStatus": {
      "enum": [
        "None",
        "Activated",
        "Deactivating",
        "Deactivated",
        "Deleted"
      ],
      "type": "string"
    },
    "ProductUpdateDto": {
      "description": "DTO to update a product",
      "type": "object",
      "properties": {
        "productCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "dustCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "menuCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "menuName": {
          "type": "string"
        },
        "vatLevel": {
          "$ref": "#/definitions/VatLevel"
        },
        "description": {
          "type": "string"
        },
        "inInventory": {
          "type": "boolean"
        },
        "percentage": {
          "format": "float",
          "type": "number"
        },
        "ascendingStock": {
          "type": "boolean"
        },
        "state": {
          "$ref": "#/definitions/RecordState"
        },
        "properties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PropertyReferenceDto"
          }
        },
        "variations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariationUpdateDto"
          }
        },
        "selectionPages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductSelectionPageReferenceDto"
          }
        }
      },
      "additionalProperties": false
    },
    "ProductVariationAddDto": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "productVariationLocations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariationLocationReferenceDto"
          }
        }
      },
      "additionalProperties": false
    },
    "ProductVariationLocationReferenceDto": {
      "type": "object",
      "properties": {
        "locationId": {
          "format": "uuid",
          "type": "string"
        },
        "unitPrice": {
          "format": "float",
          "type": "number"
        },
        "unitPricePurchase": {
          "format": "float",
          "type": "number"
        },
        "showOnFavoritePage": {
          "type": "boolean"
        },
        "minAmount": {
          "format": "float",
          "type": "number"
        },
        "maxAmount": {
          "format": "float",
          "type": "number"
        },
        "transport": {
          "format": "float",
          "type": "number"
        },
        "status": {
          "$ref": "#/definitions/ProductStatus"
        },
        "nextStatus": {
          "$ref": "#/definitions/ProductStatus"
        },
        "changeDateTime": {
          "format": "date-time",
          "type": "string"
        },
        "onMenuStart": {
          "format": "time",
          "type": "string"
        },
        "onMenuEnd": {
          "format": "time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ProductVariationUpdateDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "sku": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "productVariationLocations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariationLocationReferenceDto"
          }
        }
      },
      "additionalProperties": false
    },
    "PropertyReferenceDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "RecordState": {
      "enum": [
        "Active",
        "Inactive",
        "Deleted"
      ],
      "type": "string"
    },
    "VatLevel": {
      "enum": [
        "None",
        "Low",
        "High"
      ],
      "type": "string"
    }
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://auth.juliapos.eu/connect/authorize",
      "tokenUrl": "https://auth.juliapos.eu/connect/token",
      "scopes": {
        "openid": "openid",
        "demo": "demo"
      }
    }
  },
  "security": [
    {
      "oauth2": [ ]
    }
  ]
}